IRAD-61: renderable improvements & refactoring
Added an abstract base class for the renderables to try to start improving on them in such a way that  we can  leverage  them in a more dynamic and customized way, showing renderables with different configurations without having to duplicate the satellite list that backs them, to be specific. Tweaked what used to be a “debug  feature” that was commented out to allow toggling between showing a heatmap of 2D coverage which evolves over time vs “spotlight  mode” which just shows the current coverage  of a satellite as a simple 2D circle on the ground rather than as the 3D cone from 3D coverage. Changed the coverage renderable to extend AnalyticSurface so that I can make it clamp to ground which fixes some issues with it being a layer that hovered  a bit above the ground. Misc minor cleanup and refactoring.